backend/requirements.txt
fastapi
uvicorn
llama-cpp-python
fpdf
python-multipart
httpx
requests-mock

# Database
sqlalchemy>=2.0
psycopg2-binary
alembic

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# Validation & Configuration
pydantic>=2.0
pydantic-settings
email-validator

# Testing
pytest
pytest-asyncio
